+++ frontend/src/app/features/work-packages/components/wp-table/timeline/cells/wp-timeline-cell.ts
@@ -38,19 +38,9 @@ import { SchemaCacheService } from 'core-app/core/schemas/schema-cache.service';
 import { registerWorkPackageMouseHandler } from './wp-timeline-cell-mouse-handler';
 import { TimelineMilestoneCellRenderer } from './timeline-milestone-cell-renderer';
 import { TimelineCellRenderer } from './timeline-cell-renderer';
+import { WorkPackageCellLabels } from './wp-timeline-cell-labels';
 import { RenderInfo } from '../wp-timeline';
 import { WorkPackageTimelineTableController } from '../container/wp-timeline-container.directive';
-
-export class WorkPackageCellLabels {
-  constructor(public readonly center:HTMLDivElement|null,
-    public readonly left:HTMLDivElement,
-    public readonly leftHover:HTMLDivElement|null,
-    public readonly right:HTMLDivElement,
-    public readonly rightHover:HTMLDivElement|null,
-    public readonly farRight:HTMLDivElement,
-    public readonly withAlternativeLabels?:boolean) {
-  }
-}
 
 export class WorkPackageTimelineCell {
   @InjectField() halEditing:HalResourceEditingService;
